requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
#paho-mqtt==1.6.1
#aiosqlite==0.19.0
#sqlalchemy==2.0.23
jinja2==3.1.2
#python-multipart==0.0.6
python-dotenv
requests
psycopg[binary]